fix(phase40): Add translateTypeToCContext helper to fix cross-ASTContext type mixing
Problem:
- QualTypes from C++ ASTContext cannot be used in C ASTContext
- Previous code copied types directly, causing segfaults
- Math Library crashed during transpilation with non-constructor initializers

Solution:
- Created translateTypeToCContext() helper function that properly converts types
- Handles: builtin types, record types, pointers, const, references->pointers
- Maps C++ builtin types to equivalent C context builtins
- Uses struct names to recreate record types in C context
- Recursively translates pointer pointee types

Changes:
- include/CppToCVisitor.h: Add translateTypeToCContext declaration
- src/CppToCVisitor.cpp:
  - Implement comprehensive type translation (lines 2460-2540)
  - Replace manual type handling in VisitCXXMethodDecl (simplified)
  - Replace manual type handling in VisitCXXConstructorDecl (simplified)
  - Replace manual type handling in VisitFunctionDecl (simplified)
  - Replace manual type handling in translateCallExpr (simplified)

Results:
- Math Library now transpiles, compiles, and runs successfully (PASS)
- Pass rate improved from 0/5 (0%) to 1/5 (20%)
- Remaining failures are pre-existing bugs unrelated to this fix:
  - Custom Container/State Machine/Game Logic: Header emission bugs
  - Simple Parser: SpecialOperatorTranslator crash

+++ tests/real-world/simple-validation/02-custom-container/transpiled/main.c
@@ -3,7 +3,7 @@
 
 #include "main.h"
 
-size_t LinkedList_size(struct LinkedList * this);
+unsigned long LinkedList_size(struct LinkedList * this);
 bool LinkedList_isEmpty(struct LinkedList * this);
 void LinkedList_push_back_const_int_ref(struct LinkedList * this, const int * value);
 int LinkedList_front(struct LinkedList * this);
