+++ tests/real-world/simple-validation/02-custom-container/transpiled/main.h
@@ -12,7 +12,7 @@
 
 #include "src/LinkedList.h"
 
-size_t LinkedList_size(struct LinkedList * this);
+unsigned long LinkedList_size(struct LinkedList * this);
 bool LinkedList_isEmpty(struct LinkedList * this);
 void LinkedList_push_back_const_int_ref(struct LinkedList * this, const int * value);
 int LinkedList_front(struct LinkedList * this);
